Added antialiasing.  Changed Scene.cpp to use antialiasing as described on page 311 of tiger.  Also, commented out instances where we used new and instead just used stack memory to make the objects.  This removed memory leaks.

+++ trunk/rayTracer/Sample.cpp
@@ -4,7 +4,7 @@ Sample::Sample(void)
 {
 }
 
-Sample::Sample(int posx, int posy)
+Sample::Sample(float posx, float posy)
 {
 	x = posx; 
 	y = posy;
